INS-3076: added helper to get request from global context, refactored a bit
